requirements.txt
lxml==4.9.1
numpy==1.23.5
opencv_python==4.7.0.72
pytest==7.1.2
python-dotenv==1.0.0
Requests==2.28.2
tweepy==4.13.0
selenium==4.11.2
